Fix user api
- Delete times of user when deleting the user
- Update times when changing user name

+++ src/app/api/user/route.ts
@@ -78,22 +78,38 @@ export async function POST(request: NextRequest) {
     updateData.password = await hash(json.password, 12);
   }
 
-  const result = await prisma.user.update({
-    where: {
-      id: parseInt(json.id),
-    },
-    data: updateData,
-    select: {
-      id: true,
-      username: true,
-      email: true,
-      role: true,
-      updatedAt: true,
-      createdAt: true,
-    },
+  const user = await prisma.user.findUnique({
+    where: { id: json.id },
   });
 
-  return NextResponse.json({ result });
+  if (!user) return NextResponse.error();
+
+  const [userResult, timesResult] = await prisma.$transaction([
+    prisma.user.update({
+      where: {
+        id: parseInt(json.id),
+      },
+      data: updateData,
+      select: {
+        id: true,
+        username: true,
+        email: true,
+        role: true,
+        updatedAt: true,
+        createdAt: true,
+      },
+    }),
+    prisma.times.updateMany({
+      where: {
+        user: user.username,
+      },
+      data: {
+        user: updateData.username,
+      },
+    }),
+  ]);
+
+  return NextResponse.json({ userResult, timesResult });
 }
 
 // Delete
@@ -107,11 +123,18 @@ export async function DELETE(request: NextRequest) {
 
   if (json.id == 1) return NextResponse.error();
 
-  const result = await prisma.user.delete({
-    where: {
-      id: parseInt(json.id),
-    },
-  });
+  const userToDelete = await prisma.user.findUnique({ where: { id: json.id } });
 
-  return NextResponse.json({ result });
+  if (!userToDelete) return NextResponse.error();
+
+  const [timesResult, userResult] = await prisma.$transaction([
+    prisma.times.deleteMany({ where: { user: userToDelete.username } }),
+    prisma.user.delete({
+      where: {
+        id: userToDelete.id,
+      },
+    }),
+  ]);
+
+  return NextResponse.json({ timesResult, userResult });
 }
